requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
boto3==1.35.29
botocore==1.35.29
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cohere==5.10.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
fastapi==0.115.0
fastavro==1.9.7
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.7
lida==0.0.14
llmx==0.0.21a0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-venn==1.1.1
mdurl==0.1.2
mizani==0.11.4
narwhals==1.8.4
networkx==3.3
numpy==2.1.1
openai==1.50.2
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
patsy==0.5.6
pillow==10.4.0
plotly==5.24.1
plotnine==0.13.6
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-multipart==0.0.11
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.6
statsmodels==0.14.3
streamlit==1.38.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
watchdog==4.0.2
wordcloud==1.9.3
